Вынес общие для 2х экранов элементы интерфейса в один файл
